docs(installer): Have installer update honeybee-energy-standards
I have come to realize that all user-specific standards will be in the honeybee-standards folder and users should not be editing honeybee-energy-standards. Moreover, we will definitely be pushing updates to honeybee-energy-standards as support is added for ASHRAE 90.1 2016 and as we add SHW to the programs once it's available in the schema. So I think it's best to update honeybee-energy-standards whenever the versioner is run.

This change also makes the installer compatible with Rhino 7 if it is installed.